+++ R/calculate_diff_abundance.R
@@ -107,6 +107,11 @@ diff_abundance <-
 #' @param p_adj_method a character value, specifies the p-value correction method. Possible
 #' methods are c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). Default
 #' method is `"BH"`.
+#' @param limma_legacy_estimation a logical value, specifying if the new or old hyperparameter
+#' estimation of the limma package should be used. See the `legacy` argument of `?limma:eBayes`
+#' for more information. Our default is the same as for the `limma` package, which introduces
+#' a changed and more robust estimation of hyperparameters. This may cause changes to the
+#' results when compared to before August 2024 (limma 3.61.8).
 #' @param retain_columns a vector indicating if certain columns should be retained from the input
 #' data frame. Default is not retaining additional columns `retain_columns = NULL`. Specific
 #' columns can be retained by providing their names (not in quotations marks, just like other
@@ -214,6 +219,7 @@ calculate_diff_abundance <-
            filter_NA_missingness = TRUE,
            method = c("moderated_t-test", "t-test", "t-test_mean_sd", "proDA"),
            p_adj_method = "BH",
+           limma_legacy_estimation = NULL,
            retain_columns = NULL) {
     . <- NULL
     if (!(ref_condition %in% unique(pull(data, {{ condition }}))) & ref_condition != "all") {
@@ -553,7 +559,11 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
       message("DONE", appendLF = TRUE)
       message("[6/7] Compute empirical Bayes statistics ... ", appendLF = FALSE)
 
-      moderated_t_test_fit3 <- limma::eBayes(moderated_t_test_fit2)
+      if (packageVersion("limma") < "3.61.8"){
+        moderated_t_test_fit3 <- limma::eBayes(moderated_t_test_fit2)
+      } else {
+        moderated_t_test_fit3 <- limma::eBayes(moderated_t_test_fit2, legacy = limma_legacy_estimation)
+      }
 
       message("DONE", appendLF = TRUE)
       message("[7/7] Create result table ... ", appendLF = FALSE)
@@ -605,6 +615,10 @@ missingness type is assigned.\n The created comparisons are: \n", prefix = "\n",
         dplyr::arrange(.data$adj_pval, .data$pval)
 
       message("DONE", appendLF = TRUE)
+      # Check the limma version
+      if (packageVersion("limma") < "3.61.8" & !missing(limma_legacy_estimation)) {
+        warning("You provided the 'limma_legacy_estimation' argument, but your limma version does not support it. Update to a version higher than 3.61.8.")
+      }
 
       if (!missing(retain_columns)) {
         moderated_t_test_adj_pval <- data %>%