+++ tests/testthat/test-workflow.R
@@ -326,15 +326,8 @@ test_that("calculate_diff_abundance works", {
     expect_equal(ncol(diff_moderated), 13)
     expect_equal(ncol(diff_proDA), 12)
     expect_equal(round(min(diff_mean_sd$adj_pval, na.rm = TRUE), digits = 9), 0.00758761)
-    expect_equal(round(min(diff_moderated$adj_pval, na.rm = TRUE), digits = 9), 5.7616e-05)
+    expect_equal(round(min(diff_moderated$adj_pval, na.rm = TRUE), digits = 9), 3.87e-05)
     expect_equal(round(min(diff_proDA$adj_pval, na.rm = TRUE), digits = 5), 0.00125)
-
-    # For debugging
-    expect_equal(nrow(diff_moderated), 601)
-    expect_equal(round(min(diff_moderated$pval, na.rm = TRUE), digits = 9), 1.53e-07)
-    expect_equal(diff_moderated[1,]$peptide, "peptide_2_1")
-    expect_equal(diff_moderated %>% filter(pval < 0.05) %>% nrow(), 42)
-    expect_equal(diff_moderated %>% filter(adj_pval < 0.05) %>% nrow(), 4)
   }
 })
 
@@ -496,7 +489,7 @@ if (Sys.getenv("TEST_PROTTI") == "true") {
     expect_equal(ncol(diff_moderated_deprecated), 13)
     expect_equal(ncol(diff_proDA_deprecated), 12)
     expect_equal(round(min(diff_mean_sd_deprecated$adj_pval, na.rm = TRUE), digits = 9), 0.00758761)
-    expect_equal(round(min(diff_moderated_deprecated$adj_pval, na.rm = TRUE), digits = 9), 5.7616e-05)
+    expect_equal(round(min(diff_moderated_deprecated$adj_pval, na.rm = TRUE), digits = 9), 3.87e-05)
     expect_equal(round(min(diff_proDA_deprecated$adj_pval, na.rm = TRUE), digits = 5), 0.00125)
   })
 }